- fix for extra icons in modern - code cleaning
git-svn-id: http://svn.miranda-ng.org/main/trunk@2004 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ include/m_clistint.h
@@ -102,7 +102,7 @@ struct ClcContactBase
 	BYTE flags;
 	union {
 		struct {
-			WORD iImage;
+			int    iImage;
 			HANDLE hContact;
 		};
 		struct {
